+++ amy/headers.py
@@ -1,8 +1,8 @@
-# amy_headers.py
-# Generate headers for libAMY
+# headers.py
+# Generate headers for AMY
 import sys
-
 import numpy as np
+import amy
 
 def generate_amy_pcm_header(sample_set, name, pcm_AMY_SAMPLE_RATE=22050):
     from sf2utils.sf2parse import Sf2File
@@ -26,7 +26,7 @@ def generate_amy_pcm_header(sample_set, name, pcm_AMY_SAMPLE_RATE=22050):
         try:
             sf2 = Sf2File(open(fn, 'rb'))
         except:
-            print("For PCM presetes, download the sf2 files first. See comment in amy_headers.generate_amy_pcm_header()")
+            print("For PCM presetes, download the sf2 files first. See comment in amy.headers.generate_amy_pcm_header()")
             return
         if is_inst:
             for i,inst in enumerate(sf2.instruments[:-1]):
@@ -70,7 +70,7 @@ def generate_amy_pcm_header(sample_set, name, pcm_AMY_SAMPLE_RATE=22050):
         b.write(struct.pack('<h', all_samples[i]))
     b.close()
     p = open("src/pcm_%s.h" % (name), "w")
-    p.write("// Automatically generated by amy_headers.generate_pcm_header()\n")
+    p.write("// Automatically generated by amy.headers.generate_pcm_header()\n")
     p.write("#ifndef __PCM_H\n#define __PCM_H\n#include \"pcm_samples_%s.h\"\n" % (name))
     p.write("%sconst uint16_t pcm_samples=%d;\n#define PCM_LENGTH %d\n#define PCM_AMY_SAMPLE_RATE %d\n" % (weak, len(offsets), all_samples.shape[0], pcm_AMY_SAMPLE_RATE))
     p.write("%sconst pcm_map_t pcm_map[%d] PROGMEM = {\n" % (weak, len(offsets)))
@@ -81,7 +81,7 @@ def generate_amy_pcm_header(sample_set, name, pcm_AMY_SAMPLE_RATE=22050):
     p.close()
 
     p = open("src/pcm_samples_%s.h" % (name), 'w')
-    p.write("// Automatically generated by amy_headers.generate_pcm_header()\n")
+    p.write("// Automatically generated by amy.headers.generate_pcm_header()\n")
     p.write("#ifndef __PCM_SAMPLES_H\n#define __PCM_SAMPLES_H\n")
     p.write("%sconst int16_t pcm[%d] PROGMEM = {\n" % (weak, all_samples.shape[0]))
 
@@ -359,8 +359,7 @@ def make_piano_patch():
 def make_patches(filename):
     def nothing(message):
         return
-
-    import juno, amy, fm
+    from . import juno, fm
     num_oscs =[]
     # Don't make any noise
     amy.override_send = nothing
@@ -452,7 +451,7 @@ def make_interp_partials(filename, data_dict):
     Generate all the headers except for the partials headers
 """
 def generate_all():
-    import fm
+    from . import fm
     import collections
     import json
     # Implement the multiple lookup tables.
