+++ amy/wave.py
